requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
antigravity==0.1
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.2.25
click==8.1.8
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pytest==8.4.2
starlette==0.49.3
tomli==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
